Added some jquery functions
Trying to get events to show on weekly and daily cal. Json response is
there. Also working on pop ups and click functions. We may need to just
have a monthly. Focus and get that working, then add in functionality
for the weekly and daily stuff. Although I'll probably obsess over this
for a while.